# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "kry(Kim-gi-dong, Ryo, Yun) is simple problem-solving contest server written in Go.",
        "title": "kry-go",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Ryo",
            "url": "https://github.com/gwanryo/kry-go",
            "email": "gwanryo@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/gwanryo/kry-go/blob/main/LICENSE"
        },
        "version": "0.1"
    },
    "paths": {
        "/contest": {
            "get": {
                "description": "Get contests",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "Get contests",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Page of contests",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.ContestResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Register contest",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "Post contest",
                "parameters": [
                    {
                        "description": "Necessary contest information",
                        "name": "contest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ContestRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/contest/:contestId": {
            "get": {
                "description": "Get contest detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "Get contest detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Contest ID",
                        "name": "contest_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ContestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "description": "Edit contest detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "Put contest detail",
                "parameters": [
                    {
                        "description": "Necessary contest information",
                        "name": "contest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ContestRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete contest detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "Delete contest detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Contest ID",
                        "name": "contest_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
        "request.ContestRequest": {
            "description": "Necessary contest information when request",
            "type": "object",
            "required": [
                "content",
                "ended_at",
                "problem_ids",
                "started_at",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 191
                },
                "ended_at": {
                    "type": "string"
                },
                "problem_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "started_at": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "response.ContestResponse": {
            "description": "Contest information server responses",
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "endedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "problemIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "startedAt": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}